+++ Src/library/net/Clib/syncpoll.c
@@ -44,7 +44,7 @@ indexing
 #include <sys/ioctl.h>
 #endif
 
-#include <fcntl.h> /*x added for fcntl */
+#include <fcntl.h> /* added for fcntl */
 #ifndef FNDELAY
 #define FNDELAY O_NDELAY
 #endif
@@ -81,24 +81,24 @@ indexing
 #define nanoseconds_in_microsecond RTU64C(1000)
 
 EIF_INTEGER c_syncpoll(EIF_INTEGER fd)
-	/*x Synchronously poll a socket without modifying buffer
+	/* Synchronously poll a socket without modifying buffer
 	    expecting 1 for something, 0 for eof, -1 for error */
 {
 	char single_char_buf;
 	int result;
 
-	result = recv((int) fd, (char *) &single_char_buf, (int) 1, (int) MSG_PEEK);
+	result = recv((int) fd, (char *) &single_char_buf, 1, (int) MSG_PEEK);
 	return (EIF_INTEGER) result;
 }
 
 EIF_INTEGER c_select_poll(EIF_INTEGER fd)
-	/*x Get read status for socket fd */
+	/* Get read status for socket fd */
 {
 	fd_set fdmask;
 	struct timeval timeout;
 
-	timeout.tv_sec = (long) 0;
-	timeout.tv_usec = (long) 0;
+	timeout.tv_sec = 0;
+	timeout.tv_usec = 0;
 
 	FD_ZERO(&fdmask);
 	FD_SET(fd, &fdmask);
@@ -111,7 +111,7 @@ EIF_INTEGER c_select_poll(EIF_INTEGER fd)
 EIF_INTEGER c_select_poll_with_timeout(EIF_INTEGER fd, 
 		                               EIF_BOOLEAN read_mode,
 									   EIF_NATURAL_64 timeout_ns)
-	/*x Get read/write status for socket fd within `timeout_ns` nanoseconds*/
+	/* Get read/write status for socket fd within `timeout_ns` nanoseconds*/
 {
 	fd_set fdmask;
 	struct timeval tmout;
@@ -120,7 +120,7 @@ EIF_INTEGER c_select_poll_with_timeout(EIF_INTEGER fd,
 	n64 = timeout_ns / nanoseconds_in_second;
 	if (n64 > LONG_MAX) {
 		/* Overflown, use max value instead */
-		tmout.tv_sec = (long) LONG_MAX;
+		tmout.tv_sec = LONG_MAX;
 	} else {
 		tmout.tv_sec = (long) n64;
 	}
@@ -129,7 +129,7 @@ EIF_INTEGER c_select_poll_with_timeout(EIF_INTEGER fd,
 			/* timeout values are above zero, but less than 1 microsecond
 			 * then set to 1 microsecond to avoid eventual special behavior for 0 timeouts).
 			 */
-		tmout.tv_usec = (long) 1;
+		tmout.tv_usec = 1;
 	}
 
 	FD_ZERO(&fdmask);
@@ -147,7 +147,7 @@ EIF_INTEGER c_select_poll_with_timeout(EIF_INTEGER fd,
 
 EIF_INTEGER c_check_exception_with_timeout(EIF_INTEGER fd, 
 									       EIF_NATURAL_64 timeout_ns)
-	/*x Get exception status for socket fd within `timeout_ns` nanoseconds */
+	/* Get exception status for socket fd within `timeout_ns` nanoseconds */
 {
 	fd_set fdmask;
 	struct timeval tmout;
@@ -156,7 +156,7 @@ EIF_INTEGER c_check_exception_with_timeout(EIF_INTEGER fd,
 	n64 = timeout_ns / nanoseconds_in_second;
 	if (n64 > LONG_MAX) {
 		/* Overflown, use max value instead */
-		tmout.tv_sec = (long) LONG_MAX;
+		tmout.tv_sec = LONG_MAX;
 	} else {
 		tmout.tv_sec = (long) n64;
 	}
@@ -165,7 +165,7 @@ EIF_INTEGER c_check_exception_with_timeout(EIF_INTEGER fd,
 			/* timeout values are above zero, but less than 1 microsecond
 			 * then set to 1 microsecond to avoid eventual special behavior for 0 timeouts).
 			 */
-		tmout.tv_usec = (long) 1;
+		tmout.tv_usec = 1;
 	}
 
 	FD_ZERO(&fdmask);
@@ -179,8 +179,8 @@ EIF_INTEGER c_check_exception_with_timeout(EIF_INTEGER fd,
 }
 
 EIF_INTEGER c_is_blocking(EIF_INTEGER fd)
-	/*x attempt to get blocking status of socket */
-	/*x BIG BUG UNDER HP-UX !!! => couldn't get actual blocking status */
+	/* Attempt to get blocking status of socket */
+	/* BIG BUG UNDER HP-UX !!! => couldn't get actual blocking status */
 {
 #if defined EIF_WINDOWS || defined EIF_VMS || defined VXWORKS
 	return 0;
